backend/config/urls.py
from django.contrib import admin
from django.urls import include, path
from rest_framework.response import Response
from rest_framework.views import APIView


class HealthCheckView(APIView):
    authentication_classes = []
    permission_classes = []

    def get(self, request):
        return Response({"status": "ok"})


urlpatterns = [
    path("admin/", admin.site.urls),
    path("api/v1/health/", HealthCheckView.as_view()),
    path("api/v1/merchants/", include("apps.merchants.urls")),
    path("api/v1/payouts/", include("apps.payouts.urls")),
]