backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.4.0
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.6.0
djangorestframework==3.15.2
iniconfig==2.3.0
kombu==5.6.2
packaging==26.2
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
gunicorn==23.0.0